requirements.txt
asgiref==3.5.2
astroid==2.15.5
attrs==22.1.0
autopep8==2.0.0
aws-cfn-bootstrap==2.0
awscli==2.14.5
awscrt==0.19.19
Babel==2.9.1
boto3==1.34.69
botocore==1.34.69
certifi==2022.12.7
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.1.1
chevron==0.13.1
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
cycler==0.12.1
dbus-python==1.2.18
dill==0.3.8
distro==1.5.0
Django==2.1.15
docutils==0.16
ec2-hibinit-agent==1.0.8
fonttools==4.51.0
git-remote-codecommit==1.17
gpg==1.15.1
idna==3.4
ikp3db==1.4.2
importlib_resources==6.4.0
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.21
jsonpointer==2.0
jsonschema==4.17.3
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
libcomps==0.1.20
lockfile==0.12.2
MarkupSafe==2.1.1
matplotlib==3.8.4
mccabe==0.7.0
mercurial==5.7.1
netifaces==0.10.6
numpy==1.26.4
oauthlib==3.0.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==22.0
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.20.0
pylint==2.17.4
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.8.2
PyMySQL==1.1.0
pyotp==2.7.0
pyparsing==2.4.7
pyrsistent==0.19.2
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.9.0.post0
pytz==2022.6
PyYAML==6.0
release-notification==1.2
requests==2.28.1
rpm==4.16.1.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.16.0
sqlparse==0.4.3
support-info==1.0
swagger-spec-validator==3.0.3
systemd-python==235
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
wrapt==1.16.0
zipp==3.18.1
